Fix pdsadmin account list on Ubuntu 20.04
Currently on Ubuntu 20.04 when running pdsadmin account list the command fails with the error `column: invalid option -- '-'`. On Ubuntu 20.04 the column command only accepts the -t. Later supported versions of Ubuntu and Debian support both --table and -t. So this change updates the list command to use -t for column to support all supported distributions.