pdsadmin.sh: TOFU, require permission for command updates
Running code off the internet is not a healthy pattern. While a user might trust the repository at the time they first install the code, they should be aware of subsequent updates and have the opportunity to inspect them before running new code. This change provides a minimally intrusive prompt (defaulting to allow on hitting enter), and does not prompt on first installation.